Extract json stream processing into separate method
